+++ contracts/Endpoints.sol
@@ -35,23 +35,23 @@ import { Label } from "./commands/admin/Label.sol";
 import { Unauthorize } from "./commands/admin/Unauthorize.sol";
 
 // Port endpoints
-import { PortBase, encodePortCall } from "./ports/Base.sol";
-import { PortAllowAssets, IPortAllowAssets } from "./ports/AllowAssets.sol";
-import { PortAllowance, IPortAllowance } from "./ports/Allowance.sol";
-import { PortRedeemBalance, RedeemBalanceHook, IPortRedeemBalance } from "./ports/Redeem.sol";
-import { PortCreditAccount, IPortCreditAccount } from "./ports/Credit.sol";
-import { PortDebitAccount, IPortDebitAccount } from "./ports/Debit.sol";
-import { PortDenyAssets, IPortDenyAssets } from "./ports/DenyAssets.sol";
-import { PortPipePayable, IPortPipePayable } from "./ports/Pipe.sol";
-import { PortDispatchPayable, IPortDispatchPayable } from "./ports/Dispatch.sol";
-import { PortSettle, IPortSettle } from "./ports/Settle.sol";
+import { PortBase } from "./ports/Base.sol";
+import { PortAllowAssets } from "./ports/AllowAssets.sol";
+import { PortAllowance } from "./ports/Allowance.sol";
+import { PortRedeemBalance, RedeemBalanceHook } from "./ports/Redeem.sol";
+import { PortCreditAccount } from "./ports/Credit.sol";
+import { PortDebitAccount } from "./ports/Debit.sol";
+import { PortDenyAssets } from "./ports/DenyAssets.sol";
+import { PortPipePayable } from "./ports/Pipe.sol";
+import { PortDispatchPayable } from "./ports/Dispatch.sol";
+import { PortSettle } from "./ports/Settle.sol";
 
 // Guard endpoints
-import { GuardBase, encodeGuardCall } from "./guards/Base.sol";
+import { GuardBase } from "./guards/Base.sol";
 import { Revoke } from "./guards/Revoke.sol";
 
 // Query endpoints
-import { QueryBase, encodeQueryCall } from "./queries/Base.sol";
+import { QueryBase } from "./queries/Base.sol";
 import { AssetStatus, AssetStatusHook } from "./queries/Assets.sol";
 import { GetBalances, GetBalancesHook } from "./queries/Balances.sol";
 import { GetPosition, GetPositionHook } from "./queries/Positions.sol";