+++ contracts/ports/Settle.sol
@@ -8,14 +8,10 @@ import { Cursors, Cur, Schemas } from "../Cursors.sol";
 
 using Cursors for Cur;
 
-interface IPortSettle {
-    function portSettle(bytes calldata data) external returns (bytes memory);
-}
-
 /// @title PortSettle
 /// @notice Port that consumes peer-supplied TRANSACTION blocks through debit and credit hooks.
 /// Each TRANSACTION block calls `debitAccount` for `from` and `creditAccount` for `to`.
-abstract contract PortSettle is PortBase, DebitAccountHook, CreditAccountHook, IPortSettle {
+abstract contract PortSettle is PortBase, DebitAccountHook, CreditAccountHook {
     uint internal immutable portSettleId = portId(this.portSettle.selector);
 
     constructor() {